Try to resolve cells out-of-sync issues
1) If an instance was deleted from the API cell while still existing in a
child cell, instance updates would fail at the API cell.. and even
forced syncing would fail.  Solution is to use read_deleted='yes' when updating.

2) Race conditions on delete.  One can do a delete on an instance in the
API cell before it knows what cell the instance landed in.  Before this
patch, the instance would be removed from the API cell.. and would still
build in a child cell.  And while the instance would be deleted in the
API cell, a 404 would also be raised.  This is a case that would cause
1) above.  Along with the fix for #1, the other part is to not return a
404.. and also try to broadcast a destroy down to all cells.  Worst
case, if there's still a race between the build+destroy in the child,
an instance_update would now revive the instance in the API cell due to
1) being fixed.

3) Catch conditions where the API tries to do an action on an instance
that is found to be deleted in a child cell.  Have the child cell tell
the parents that the instance is gone.