Remove bandwidth data from immediate exists events.
When we were polling XenServer RRD's for bandwidth, it would zero the
bandwidth used on resize/rebuilds. Since nova itself is now doing the
polling of the VIFs, this no longer happens.

Usage was previously adding together bandwidth from the immediate .exists
events emitted on resize/rebuilds with the value from the end-of-day events
to compensate. That is no longer needed, so we need to remove the bandwidth
data from the immediate events to prevent overbilling.